define useAssignedLevel for use in ClassifiyPageContainer too
